Deposit plugins: extract models.ts + reorder classes; clean unused imports
Refactor (no behavior change):
  - OTA: OtaIntent + OtaResult moved to ota-deposit/models.ts
  - Pull: PullIntent + PullResult moved to pull-deposit/models.ts (out of
    approval-watcher.ts)
  - OtaDepositPlugin and PullDepositPlugin classes reordered so the public
    PaymentsPlugin interface methods (deposit, depositCustom, payout) sit at
    the top of the class, helpers below — top-to-bottom now reads contract
    first, mechanism second.

Dead code:
  - wallet-deposit/plugin.ts: dropped unused `logger` field (and the winston
    import) — the class never logs internally
  - fireblocks/config.ts: dropped unused `import * as fs` (stale since #215)
  - src/index.ts: dropped unused FinP2PContract + InMemoryExecDetailsStore
    imports
