+++ src/index.ts
@@ -2,12 +2,10 @@ import * as process from "process";
 import { logger } from "@owneraio/finp2p-nodejs-skeleton-adapter";
 import { FinP2PClient } from "@owneraio/finp2p-client";
 import winston, { format, transports } from "winston";
-import { FinP2PContract } from "@owneraio/finp2p-contracts";
 import { migrationsDir as vanillaMigrationsDir, migrationsTableName as vanillaMigrationsTable } from "@owneraio/finp2p-vanilla-service";
 import { join } from "path";
 import { envVarsToAppConfig } from "./config";
 import createApp from "./app";
-import { InMemoryExecDetailsStore } from "./services/finp2p-contract";
 
 const init = async () => {
   const port = process.env.PORT || "3000";
